mock-nws-api/requirements.txt
azure-functions
requests
